✨ abstraction ✨
